+++ src/Dca/DcaUtil.php
@@ -25,7 +25,9 @@ use Contao\StringUtil;
 use Contao\System;
 use Contao\Validator;
 use Doctrine\DBAL\Connection;
+use HeimrichHannot\UtilsBundle\Choice\ModelInstanceChoice;
 use HeimrichHannot\UtilsBundle\Driver\DC_Table_Utils;
+use HeimrichHannot\UtilsBundle\Model\ModelUtil;
 use HeimrichHannot\UtilsBundle\Routing\RoutingUtil;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 
@@ -38,6 +40,7 @@ class DcaUtil
     const AUTHOR_TYPE_NONE = 'none';
     const AUTHOR_TYPE_MEMBER = 'member';
     const AUTHOR_TYPE_USER = 'user';
+    const AUTHOR_TYPE_SESSION = 'session';
 
     /** @var ContaoFrameworkInterface */
     protected $framework;
@@ -745,8 +748,8 @@ class DcaUtil
         $this->framework->getAdapter(Controller::class)->loadDataContainer($table);
 
         // callbacks
-        $GLOBALS['TL_DCA'][$table]['config']['oncreate_callback']['setAuthorIDOnCreate'] = ['huh.utils.dca', 'setAuthorIDOnCreate'];
-        $GLOBALS['TL_DCA'][$table]['config']['onload_callback']['modifyAuthorPaletteOnLoad'] = ['huh.utils.dca', 'modifyAuthorPaletteOnLoad', true];
+        $GLOBALS['TL_DCA'][$table]['config']['oncreate_callback']['setAuthorIDOnCreate'] = [self::class, 'setAuthorIDOnCreate'];
+        $GLOBALS['TL_DCA'][$table]['config']['onload_callback']['modifyAuthorPaletteOnLoad'] = [self::class, 'modifyAuthorPaletteOnLoad', true];
 
         // fields
         $GLOBALS['TL_DCA'][$table]['fields'][$fieldPrefix ? $fieldPrefix.ucfirst(static::PROPERTY_AUTHOR_TYPE) : static::PROPERTY_AUTHOR_TYPE] = [
@@ -759,6 +762,7 @@ class DcaUtil
                 static::AUTHOR_TYPE_NONE,
                 static::AUTHOR_TYPE_MEMBER,
                 static::AUTHOR_TYPE_USER,
+                // session is only added if it's already set in the dca
             ],
             'reference' => $GLOBALS['TL_LANG']['MSC']['utilsBundle']['authorType'],
             'eval' => ['doNotCopy' => true, 'submitOnChange' => true, 'mandatory' => true, 'tl_class' => 'w50 clr'],
@@ -783,13 +787,13 @@ class DcaUtil
                 'includeBlankOption' => true,
                 'tl_class' => 'w50',
             ],
-            'sql' => "int(10) unsigned NOT NULL default '0'",
+            'sql' => "varchar(64) NOT NULL default ''",
         ];
     }
 
     public function setAuthorIDOnCreate(string $table, int $id, array $row, DataContainer $dc)
     {
-        $model = $this->container->get('huh.utils.model')->findModelInstanceByPk($table, $id);
+        $model = $this->container->get(ModelUtil::class)->findModelInstanceByPk($table, $id);
         /** @var Database $db */
         $db = $this->framework->createInstance(Database::class);
 
@@ -803,6 +807,11 @@ class DcaUtil
             if (FE_USER_LOGGED_IN) {
                 $model->{static::PROPERTY_AUTHOR_TYPE} = static::AUTHOR_TYPE_MEMBER;
                 $model->{static::PROPERTY_AUTHOR} = $this->framework->getAdapter(FrontendUser::class)->getInstance()->id;
+                $model->save();
+            } else {
+                // php session
+                $model->{static::PROPERTY_AUTHOR_TYPE} = static::AUTHOR_TYPE_SESSION;
+                $model->{static::PROPERTY_AUTHOR} = session_id();
                 $model->save();
             }
         } else {
@@ -830,16 +839,28 @@ class DcaUtil
 
         // author handling
         if ($model->{static::PROPERTY_AUTHOR_TYPE} == static::AUTHOR_TYPE_NONE) {
-            unset($dca['fields']['author']);
+            unset($dca['fields'][static::PROPERTY_AUTHOR]);
         }
 
         if ($model->{static::PROPERTY_AUTHOR_TYPE} == static::AUTHOR_TYPE_USER) {
-            $dca['fields']['author']['options_callback'] = function () {
-                return $this->container->get('huh.utils.choice.model_instance')->getCachedChoices([
+            $dca['fields'][static::PROPERTY_AUTHOR]['options_callback'] = function () {
+                return $this->container->get(ModelInstanceChoice::class)->getCachedChoices([
                     'dataContainer' => 'tl_user',
                     'labelPattern' => '%name% (ID %id%)',
                 ]);
             };
+        }
+
+        if ($model->{static::PROPERTY_AUTHOR_TYPE} == static::AUTHOR_TYPE_SESSION) {
+            $dca['fields'][static::PROPERTY_AUTHOR_TYPE]['options'] = array_merge($dca['fields'][static::PROPERTY_AUTHOR_TYPE]['options'], [static::AUTHOR_TYPE_SESSION]);
+            // do not allow to edit in backend
+            $dca['fields'][static::PROPERTY_AUTHOR_TYPE]['eval']['disabled'] = true;
+
+            unset($dca['fields'][static::PROPERTY_AUTHOR]['options_callback']);
+            $dca['fields'][static::PROPERTY_AUTHOR]['inputType'] = 'text';
+            // do not allow to edit in backend
+            $dca['fields'][static::PROPERTY_AUTHOR]['eval']['disabled'] = true;
+            $dca['fields'][static::PROPERTY_AUTHOR]['label'][0] = $GLOBALS['TL_LANG']['MSC']['utilsBundle'][static::PROPERTY_AUTHOR_TYPE][self::AUTHOR_TYPE_SESSION];
         }
     }